+++ core/src/it/scala/org/locationtech/rasterframes/ref/RasterRefIT.scala
@@ -30,18 +30,18 @@ import org.locationtech.rasterframes.expressions.aggregates.TileRasterizerAggreg
 
 class RasterRefIT extends TestEnvironment {
   describe("practical subregion reads") {
-    ignore("should construct a natural color composite") {
+    it("should construct a natural color composite") {
       import spark.implicits._
       def scene(idx: Int) = URI.create(s"https://landsat-pds.s3.us-west-2.amazonaws.com" +
         s"/c1/L8/176/039/LC08_L1TP_176039_20190703_20190718_01_T1/LC08_L1TP_176039_20190703_20190718_01_T1_B$idx.TIF")
 
-      val redScene = RasterSource(scene(4))
+      val redScene = RFRasterSource(scene(4))
       // [west, south, east, north]
       val area = Extent(31.115, 29.963, 31.148, 29.99).reproject(LatLng, redScene.crs)
 
       val red = RasterRef(redScene, 0, Some(area), None)
-      val green = RasterRef(RasterSource(scene(3)), 0, Some(area), None)
-      val blue = RasterRef(RasterSource(scene(2)), 0, Some(area), None)
+      val green = RasterRef(RFRasterSource(scene(3)), 0, Some(area), None)
+      val blue = RasterRef(RFRasterSource(scene(2)), 0, Some(area), None)
 
       val rf = Seq((red, green, blue)).toDF("red", "green", "blue")
       val df = rf.select(
@@ -55,11 +55,11 @@ class RasterRefIT extends TestEnvironment {
         stats.get.dataCells shouldBe > (1000L)
       }
       
-      //import geotrellis.raster.io.geotiff.{GeoTiffOptions, MultibandGeoTiff, Tiled}
-      //import geotrellis.raster.io.geotiff.compression.{DeflateCompression, NoCompression}
-      //import geotrellis.raster.io.geotiff.tags.codes.ColorSpace
-      //val tiffOptions = GeoTiffOptions(Tiled,  DeflateCompression, ColorSpace.RGB)
-      //MultibandGeoTiff(raster, raster.crs, tiffOptions).write("target/composite.tif")
+      import geotrellis.raster.io.geotiff.compression.DeflateCompression
+      import geotrellis.raster.io.geotiff.tags.codes.ColorSpace
+      import geotrellis.raster.io.geotiff.{GeoTiffOptions, MultibandGeoTiff, Tiled}
+      val tiffOptions = GeoTiffOptions(Tiled,  DeflateCompression, ColorSpace.RGB)
+      MultibandGeoTiff(raster.raster, raster.crs, tiffOptions).write("target/composite.tif")
     }
   }
 }